better syntax and cleanup
