+++ src/main/kotlin/io/fritz2/dom/html/elements.kt
@@ -1,6 +1,5 @@
 package io.fritz2.dom.html
 
-import io.fritz2.binding.Slot
 import io.fritz2.dom.AttributeDelegate
 import io.fritz2.dom.Tag
 import io.fritz2.dom.WithText
@@ -8,7 +7,6 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.Flow
 import org.w3c.dom.Element
-import org.w3c.dom.events.MouseEvent
 
 @ExperimentalCoroutinesApi
 @FlowPreview
@@ -19,7 +17,7 @@ class Div(): Tag("div"), WithText<Element>
 @FlowPreview
 class Button(): Tag("button"), WithText<Element> {
     //TODO: structure attributes and events in interfaces
-    var onClick: Slot<MouseEvent> by Click.delegate
+    val clicks by lazy { event(Click) }
 }
 
 @ExperimentalCoroutinesApi
@@ -27,7 +25,7 @@ class Button(): Tag("button"), WithText<Element> {
 class Input(): Tag("input") {
     var value: Flow<String> by AttributeDelegate
 
-    var onChange: Slot<String> by Change.delegate
+    var changes by lazy { event(Change) }
 }
 
 @ExperimentalCoroutinesApi
