+++ src/main/kotlin/io/fritz2/dom/html/events.kt
@@ -1,31 +1,15 @@
 package io.fritz2.dom.html
 
-import io.fritz2.binding.Slot
-import io.fritz2.dom.Tag
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.FlowPreview
 import org.w3c.dom.HTMLInputElement
 import org.w3c.dom.events.Event
 import org.w3c.dom.events.MouseEvent
-import kotlin.reflect.KProperty
-
-@ExperimentalCoroutinesApi
-@FlowPreview
-interface EventDelegate<T> {
-    operator fun getValue(thisRef: Tag, property: KProperty<*>): Slot<T> = throw NotImplementedError()
-    operator fun setValue(thisRef: Tag, property: KProperty<*>, slot: Slot<T>)
-}
 
 @ExperimentalCoroutinesApi
 @FlowPreview
 open class EventType<T>(val name: String) {
     open fun extract(event: Event): T = event.unsafeCast<T>()
-
-    val delegate: EventDelegate<T> = object : EventDelegate<T> {
-        override operator fun setValue(thisRef: Tag, property: KProperty<*>, slot: Slot<T>) {
-            slot.connect(thisRef.event(this@EventType))
-        }
-    }
 }
 
 @ExperimentalCoroutinesApi
