+++ src/main/kotlin/oldDemo.kt
@@ -1,12 +1,12 @@
 import io.fritz2.binding.Patch
 import io.fritz2.binding.Seq
-import io.fritz2.binding.Var
 import io.fritz2.dom.html.html
 import io.fritz2.dom.mount
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.map
 
+/*
 @ExperimentalCoroutinesApi
 @FlowPreview
 fun oldMain() {
@@ -77,3 +77,5 @@ fun oldMain() {
         s.set(listOf("a","e","c"))
     }
 }
+
+ */
